Simplify code by removing RelationshipItemHelper
