+++ community/kernel/src/main/java/org/neo4j/kernel/impl/api/cursor/TxAbstractRelationshipCursor.java
@@ -24,7 +24,6 @@ import java.util.function.Consumer;
 import org.neo4j.cursor.Cursor;
 import org.neo4j.kernel.api.StatementConstants;
 import org.neo4j.kernel.api.cursor.EntityItemHelper;
-import org.neo4j.kernel.api.cursor.RelationshipItemHelper;
 import org.neo4j.kernel.api.txstate.TransactionState;
 import org.neo4j.kernel.impl.api.RelationshipVisitor;
 import org.neo4j.kernel.impl.util.Cursors;
@@ -35,8 +34,8 @@ import org.neo4j.storageengine.api.txstate.RelationshipState;
 /**
  * Overlays transaction state on a {@link RelationshipItem} cursor.
  */
-public abstract class TxAbstractRelationshipCursor extends RelationshipItemHelper
-        implements Cursor<RelationshipItem>, RelationshipVisitor<RuntimeException>
+public abstract class TxAbstractRelationshipCursor extends EntityItemHelper
+        implements Cursor<RelationshipItem>, RelationshipVisitor<RuntimeException>, RelationshipItem
 {
     protected final TransactionState state;
     private final Consumer<TxAbstractRelationshipCursor> instanceCache;
