+++ community/kernel/src/test/java/org/neo4j/kernel/impl/api/state/StubCursors.java
@@ -26,9 +26,10 @@ import org.neo4j.collection.primitive.PrimitiveIntCollections;
 import org.neo4j.collection.primitive.PrimitiveIntSet;
 import org.neo4j.cursor.Cursor;
 import org.neo4j.helpers.collection.Iterables;
-import org.neo4j.kernel.api.cursor.RelationshipItemHelper;
+import org.neo4j.kernel.api.cursor.EntityItemHelper;
 import org.neo4j.kernel.api.properties.DefinedProperty;
 import org.neo4j.kernel.impl.locking.Lock;
+import org.neo4j.kernel.impl.util.Cursors;
 import org.neo4j.storageengine.api.NodeItem;
 import org.neo4j.storageengine.api.PropertyItem;
 import org.neo4j.storageengine.api.RelationshipItem;
@@ -134,10 +135,71 @@ public class StubCursors
         }
     }
 
+    private abstract static class TestRelationshipItem extends EntityItemHelper implements RelationshipItem
+    {
+
+    }
+
+    public static RelationshipItem relationship( long id, int type, long start, long end )
+    {
+        return new TestRelationshipItem()
+        {
+            @Override
+            public Cursor<PropertyItem> property( int propertyKeyId )
+            {
+                return Cursors.empty();
+            }
+
+            @Override
+            public Cursor<PropertyItem> properties()
+            {
+                return Cursors.empty();
+            }
+
+            @Override
+            public long id()
+            {
+                return id;
+            }
+
+            @Override
+            public int type()
+            {
+                return type;
+            }
+
+            @Override
+            public long startNode()
+            {
+                return start;
+            }
+
+            @Override
+            public long otherNode( long nodeId )
+            {
+                if ( nodeId == start )
+                {
+                    return end;
+                }
+                else if ( nodeId == end )
+                {
+                    return start;
+                }
+                throw new IllegalStateException();
+            }
+
+            @Override
+            public long endNode()
+            {
+                return end;
+            }
+        };
+    }
+
     public static Cursor<RelationshipItem> asRelationshipCursor( final long relId, final int type,
             final long startNode, final long endNode, final Cursor<PropertyItem> propertyCursor )
     {
-        return cursor( new RelationshipItemHelper()
+        return cursor( new TestRelationshipItem()
         {
             @Override
             public long id()
